+++ src/Authentication/BaseFirewall.php
@@ -10,7 +10,6 @@ use Orisai\Auth\Authentication\Data\ExpiredLogin;
 use Orisai\Auth\Authentication\Data\Logins;
 use Orisai\Auth\Authentication\Exception\IdentityExpired;
 use Orisai\Auth\Authentication\Exception\NotLoggedIn;
-use Orisai\Auth\Authorization\AccessEntry;
 use Orisai\Auth\Authorization\Authorizer;
 use Orisai\Auth\Authorization\CurrentUserPolicyContextCreator;
 use Orisai\Clock\SystemClock;
@@ -197,13 +196,13 @@ abstract class BaseFirewall implements Firewall
 		return $identity->hasRole($role);
 	}
 
-	public function isAllowed(string $privilege, ?object $requirements = null, ?AccessEntry &$entry = null): bool
+	public function isAllowed(string $privilege, ?object $requirements = null, ?array &$entries = null): bool
 	{
 		return $this->authorizer->isAllowed(
 			$this->fetchIdentity(),
 			$privilege,
 			$requirements,
-			$entry,
+			$entries,
 			$this->contextCreator,
 		);
 	}